[ofono-config-ril-binder] Add ofono-config-ril-binder-single-sim-sim2 sub-package
This sub-packages enables the use of the second SIM slot disabling the first SIM slot.
It is useful for UIs like phosh, that currently support only one SIM card, running on devices that have SIM1 as a shared SIM/SD-slot and SIM2 can be used freely without blocking other functionality.